Fix OpenShiftPostgresqlTransactionGeneralUsageIT when run with img from RH registry
