+++ sql-db/narayana-transactions/src/test/java/io/quarkus/ts/transactions/PostgresqlTransactionGeneralUsageIT.java
@@ -5,32 +5,21 @@ import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.QuarkusScenario;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.ts.transactions.recovery.TransactionExecutor;
 
 @QuarkusScenario
-public class PostgresqlTransactionGeneralUsageIT extends TransactionCommons {
+public class PostgresqlTransactionGeneralUsageIT extends AbstractPostgresqlTransactionGeneralUsageIT {
 
     private static final String ENABLE_PREPARED_TRANSACTIONS = "--max_prepared_transactions=100";
-    static final int POSTGRESQL_PORT = 5432;
 
     @Container(image = "${postgresql.latest.image}", port = POSTGRESQL_PORT, expectedLog = "listening on IPv4 address", command = ENABLE_PREPARED_TRANSACTIONS)
     static final PostgresqlService database = new PostgresqlService().withProperty("PGDATA", "/tmp/psql");
 
     @QuarkusApplication
-    public static final RestService app = new RestService().withProperties("postgresql.properties")
-            .withProperty("quarkus.otel.exporter.otlp.traces.endpoint", jaeger::getCollectorUrl)
-            .withProperty("quarkus.datasource.jdbc.telemetry", "true")
-            .withProperty("quarkus.datasource.username", database.getUser())
-            .withProperty("quarkus.datasource.password", database.getPassword())
-            .withProperty("quarkus.datasource.jdbc.url", database::getJdbcUrl);
+    public static final RestService app = createQuarkusApp(database);
 
     @Override
     protected RestService getApp() {
         return app;
     }
 
-    @Override
-    protected TransactionExecutor getTransactionExecutorUsedForRecovery() {
-        return TransactionExecutor.TRANSACTIONAL_ANNOTATION;
-    }
 }
